move some common functions used by both switch-box and scatter-gather code to utils
